+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+- Fixes searching `RangeTree` with an empty range in it. Empty ranges are now ignored.
+
 ## 1.7.0 (Aug 5, 2016)
 - Throws `RangeError` if the given bounds are not valid (not of the following: `[]`, `()`, `[)`, `(]`).  
   Thanks, [Nikhil Benesch](https://github.com/benesch), for the hint!